Release VynCo .NET SDK v3.0.0 — aligned with Rust SDK v2.2.0
BREAKING: base URL changed from api.vynco.ch to vynco.ch/api,
Dashboard/Credit/AI models updated to match production API.

17 new endpoints (69→86): Companies (GetFull, Classification,
Structure, Acquisitions, Notes CRUD, Tags CRUD, AllTags,
ExportExcel), Persons (Search, Get), Dossiers (Generate),
Teams (Join). Company model expanded to 39 fields.

Retry logic now falls back to X-RateLimit-Reset when Retry-After
is absent, capped at 60s. Fixed culture-sensitive double formatting
in query strings. Optimized ExtractList to avoid double JSON parse.

+++ src/VynCo/Models/Credits.cs
@@ -7,7 +7,7 @@ public class CreditBalance
 {
     [JsonPropertyName("balance")] public int Balance { get; set; }
     [JsonPropertyName("monthlyCredits")] public long MonthlyCredits { get; set; }
-    [JsonPropertyName("usedThisMonth")] public int UsedThisMonth { get; set; }
+    [JsonPropertyName("usedThisMonth")] public long UsedThisMonth { get; set; }
     [JsonPropertyName("tier")] public string Tier { get; set; } = "";
     [JsonPropertyName("overageRate")] public double OverageRate { get; set; }
 }
